doc: clarify is-sandbox-enabled function semantics
The function actually tests the usage of a sandbox image. This is an
important distinction because it won't return true in case of the slim
sandbox build mode. Likewise, it will return false in the dev- or
strict-sandbox modes when packages have no available sandbox image even
though isolation is applied.