utils: fall back to 'spawn' process pool on WSL1
On Windows WSL1, the 'forkserver' method does not work because UNIX
domain sockets are not fully implemented. This manifests itself by an
EOFError because the forked server crashes during initialization. If
that happens we fall back to the 'spawn' method that is supposed to
work, even though its slow.

Fixes #562.