layers: put SCM backed layers in build tree
Usually, the project tree (where the recipes are located) and the build
tree (dev/work directory) are in the same directory. But the build
directory can be separated by using "bob init".

In this case, layers that are backed by an SCM, should go to the build
directory. There are two main reasons for this:

  1. There can be multiple build directories per project directory. Each
     build directory is managed individually and thus the layers can be
     updated independently from each other. Doing that in the common
     project directory would create conflicts.
  2. The state of the layers is stored in the build directory.
     Consequently, there is no common state that would describe the
     layer SCMs in the project tree.

Fixes #590.