input: add managedLayers policy
Historically, layers could be nested where they would form a tree
structure. This has been removed with the introduction of managed
layers in f70beb1f. Turns out, this was a bit optimistic because there
are existing projects that rely on this.

To bring back the old behaviour, add the managedLayers policy. If the
policy is on the old behaviour, nested layers are supported. To keep
things simple, any managed layer will be rejected in this case. If the
managedLayers policy is set to the new behaviour, all layers will be
placed flat in the project- resp. build-directory.

Fixes #595.