git: improve audit of unknown git modules
When auditing recipes, the potential git checkout was not performed by
Bob. Hence, no information is available if submodules shall be checked
out or not. In this case, assume that potentially present submodules
ought to be checked out recursively. Previously, we assumed that
submodules must not be checked out which is a questionable assumption.