input: tighten layer name rules
Internally, the empty string is used for the main recipes.  Allowing
path separators in the layer name will probably also blow up. Rule out
both.

Some of the type checks were also missing. Make sure that the layer name
is a string and that a layers entry is a dict if not already a string.