Add pyproject.toml
When installing Bob, sphinx-build must be available to build the
manpages. There used to be a setuptools integration but that was removed
upstream and consequently dropped here in 1346895a.  Using the
'setup_requires' key in setup.py does not work either.

Instead, add a pyproject.toml file that declares this dependency. That
seems to work for current pip versions. The only drawback is that we now
require Windows user to also temporarily install Sphinx during
installation, even though it will not be used so far.