git: implement unshallow
If the repository was initially cloned shallow, the next 'git fetch'
will not undo that. Consequently, the user may be left wondering why the
commit is not fetched even though the shallow property was removed. To
fix that, Bob must pass the --unshallow option in this case.

Unfortunately it is not allowed by git to specify --unshallow
unconditionally. So always check if the repository is shallow and only
add --unshallow in this case.