input: add support for global aliases
So far, Bob supports string substitution of dependency names. This lets
recipes decide the actual dependency at recipe parsing time. That is a
local decision of the recipe, though. Sometimes, a common indirection
for many dependencies is required.

Global recipe aliases provide such an indirection. They are configured
in the newly recognized "aliases" directory. The alias name is derived
from the YAML file name, just like for recipes and classes. If it is
just a single alias, the file only contains the target recipe name.
Optionally, several aliases can be defined by using a "multiPackage" in
the file.

The alias target is subject to string substitution.