build: use recipe package name as path name
With the advent of aliases, the package name may no longer resemble the
underlying recipe. This will be confusing to the user when trying to
find the workspaces for a particular recipe. Hence we should always use
the original package name based on the recipe instead of the possibly
aliased package name.