input: add per-dependency aliases
Each dependency of a recipe has to have a unique name. This is a problem
if two distinct variants of the same dependency need to be used, though.
To enable that feature, every dependency can optionally get an alias
name. Multiple references to the same dependency can be created, given
that each gets its own unique alias.

If the dependency is provided via 'provideDeps', the alias name will be
used for matching and in the downstream package dependencies. On the
other hand, the alias is ignored for recipe cycle analysis. This means
that aliases cannot be used to break cyclic recipe dependencies.

Fixes #314.