input: fail on unexpected indeterministic checkouts
Bob handles situations where the checkout step hash was predicted but
the actual checkout yielded another result by restarting the build.
While this case may indeed happen for indeterministic checkouts, it is
an error if it happens for officially stable checkouts.

So instead of restarting the build, give an error message. The user
should probably fix the recipe. Halting the build at this stage makes
the analysis easier.