url: always extract if workspace or subdir was created
Since the introduction of the urlScmSeparateDownload policy, the archive
and its extraction canary are not part of the workspace anymore. This
has the unfortunate consequence that we do not automatically detect if
the workspace was deleted by the user. Because the canary and the
archive are outside of the workspace, another extraction is not
triggered even though this is expected.

Fix this regression by forcing the extraction if Bob had to re-create
the workspace or the subdirectory in the workspace. This still does not
detect if the workspace was just emptied but it could be argued that
this is desired behaviour.