builder: verify extracted artifact against its audit trail
The audit trail names the expected hash of the content. Assert that the
hash sum matches after the artifact has been extracted. This should
detect any accidental corruptions of the archive.

Note that this will not help against malicious modifications. To detect
such tampering, some kind of cryptographically strong signature check
needs to be implemented.

Fixes #636.