input: add facility to add new internal string functions
Unfortunately we missed the opportunity to reserve a namespace for Bob
internal string functions. Adding new ones thus has the potential to
break existing plugins.

New functions are now added as a last step in the parsing process. If a
plugin has already declared a function with the same name, issue a
warning. The plugin function will be kept to not break existing projects.