input: add attribute to enable package dependencies
So far, dependencies named in the "depends" section are only available
during the build step. Sometimes recipes only need to gather results
from other packages. For this case, it is nice to have the dependencies
available in the package step too.

The "packageDepends" boolean attribute optionally enables access to all
dependencies in the package step.

Fixes #645.