url: fix repeated invocation for gz/xz files
If the download is not an archive but a single gz/xz compressed file,
the extractor will output the result to stdout. In this case the output
file must be opened only after we are sure that the extraction actually
happens. Previously, the output file was opened for writing
unconditionally. This truncated the file and would leave it in this
state if the extraction was skipped.