archive: treat more socket errors as warnings
When downloading from an archive, the build should not fail because of
network errors. After all, the binary archive is just a cache that might
not provide an archive. We now treat any errors up to and including the
connection establishment as warnings. Previously, refused connections
were already treated that way.

For uploads the behaviour does not change. Any network errors are
treated as errors for uploads.

Fixes #660.