archive: add strictdownload flag
By default, downloads from archives ignore most errors. Usually, this is
desired because this is just a different reason why no artifact is
found. But in CI environments, unreachable servers or similar problems
should never happen. For that, the "strictdownload" flag indicates that
the server must be reachable and respond properly. In any case, it is
never an error if a download does not find the requested archive, even
with the strictdownload flag.

See #660.