invoker: don't check for sandbox usability on Windows
When the user executes "bob build" on Windows, the sandbox will still be
implicitly enabled. Almost certainly, the recipes won't define a sandbox
image on Windows so this is completely transparent. Instead of failing
here eagerly, wait until a real sandbox image is used. In that case, the
user will get a proper error message and we don't need to fail here
eagerly.

Fixes: 59170fb6 ("build: check availability of user namespaces")