archive: fix soft errors on Windows
Apparently, there are multiple names for socket errors in the errno
module. For example, the ConnectionRefusedError is declared as
errno.ECONNREFUSED, as well as errno.WSAECONNREFUSED. The back-mapping
in errno.errorcode prefers the WSA-prefix, though, making it
incompatible with the Linux version.

Instead of using the error name, use the errno number directly.

Fixes: 58e0ce3c4 ("archive: treat more socket errors as warnings")