build: unify jobserver configuration into dedicated class
There are two aspects to the jobserver: the number of jobs and the
interaction with the jobserver though file descriptors. The parsing of
the external jobserver configuration, the generation of the MAKEFLAGS
and the handling of the job slots was scattered across all the layers.

Unify this all into a dedicated JobserverConfig class and pass an
instance of this class through the layers. It is used as input
configuration (parsing the external MAKEFLAGS), as well as the sub-make
configuration (as passed to invoked scripts).