archive: don't ignore archive management errors
The calls for maintaining an binary archive should not be linked to the
"nofail" option. This option is for ignoring upload errors. But when
managing binary artifacts, any error should be reported. Even worse,
ignoring the errors could lead to an inconsistent state, e.g., if the
deletion of an artifact failed but it still is removed from the local
database.