archive: simplify download error handling
Instead of looking at the errno of the underlying error, use the stage
of the download process as a criteria if the error should be fatal or
can be ignored. There are two stages that can be distinguished: first,
locating the artifact and, secondly, reading and extracting the
artifact. It was always the intention to ignore errors in the first
stage and fail in the second state. This is now directly visible in the
code and expressed much more explicitly.