input: define checkout-/build-/packageAuditFiles
Using these keys, selected files can be included into the audit trail.
This can be used to store any supplementary information in the audit
trail as necessary. Examples are hash sums of individual source files,
custom license texts or other build time meta data.

By default, the files are read with the UTF-8 codec. This can be changed
by setting an "encoding" attribute. Sometimes the presence of files
depends on other factors. The "if" property allows to include them
conditionally.